Fixed more TS errors
