+++ package.json
@@ -10,12 +10,14 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^29.5.3",
     "lodash.debounce": "^4.0.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.4.2",
     "react-scripts": "5.0.1",
     "recharts": "^2.3.2",
+    "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
@@ -43,13 +45,13 @@
     ]
   },
   "devDependencies": {
-    "@types/node": "^20.4.0",
-    "@types/react": "^18.2.14",
-    "@types/react-dom": "^18.2.6",
+    "@types/node": "^20.4.2",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "@types/react-router-dom": "^5.3.3",
     "autoprefixer": "^10.4.14",
     "postcss": "^8.4.24",
     "tailwind-scrollbar": "^2.1.0-preview.0",
-    "tailwindcss": "^3.3.2",
-    "typescript": "^5.1.6"
+    "tailwindcss": "^3.3.2"
   }
 }
