--- src/components/Navigation.js
@@ -1,55 +1,0 @@
-import React from "react";
-import { NavLink } from "react-router-dom";
-
-const Navigation = () => {
-	return (
-		<nav
-			className="w-fit sm:w-[45%] mt-20 flex  align-middle
-    border border-cyan rounded-lg
-    "
-		>
-			<NavLink
-				to="/"
-				end
-				className={({ isActive }) => {
-					return ` w-1/3 px-2 py-1 text-base text-center font-nunito m-2.5 ${isActive ? "bg-cyan text-gray-300"
-						: "bg-gray-200 text-gray-100 hover:text-cyan active:bg-cyan active:text-gray-300"} border-0 cursor-pointer rounded capitalize font-semibold`;
-				}}
-			>
-				Crypto
-			</NavLink>
-
-			<NavLink
-				to="/trending"
-				className={({ isActive }) => {
-					return `w-1/3 px-2 py-1 text-base text-center font-nunito m-2.5
-
-${isActive
-							? "bg-cyan text-gray-300"
-							: "bg-gray-200 text-gray-100 hover:text-cyan active:bg-cyan active:text-gray-300"
-						}
-    border-0 cursor-pointer rounded capitalize font-semibold`;
-				}}
-			>
-				trending
-			</NavLink>
-
-			<NavLink
-				to="/saved"
-				className={({ isActive }) => {
-					return `w-1/3 px-2 py-1 text-base text-center font-nunito m-2.5
-
-${isActive
-							? "bg-cyan text-gray-300"
-							: "bg-gray-200 text-gray-100 hover:text-cyan active:bg-cyan active:text-gray-300"
-						}
-    border-0 cursor-pointer rounded capitalize font-semibold`;
-				}}
-			>
-				saved
-			</NavLink>
-		</nav>
-	);
-};
-
-export default Navigation;
