--- src/pages/Crypto.js
@@ -1,16 +1,0 @@
-import React from "react";
-import { Outlet } from "react-router-dom";
-import Filters from "../components/Filters";
-import TableComponent from "../components/TableComponent";
-
-const Crypto = () => {
-  return (
-    <section className="w-11/12 sm:w-[80%] h-full flex flex-col mt-16 mb-10 sm:mb-24 relative">
-      <Filters />
-      <TableComponent />
-      <Outlet />
-    </section>
-  );
-};
-
-export default Crypto;
